app/requirements.txt
playwright <= 1.44.0
bs4 <= 0.0.2
boto3 <= 1.34.137
pyyaml <= 6.0.1